"/Makefile" & "/compile.sh": much improved the build-time output, improved the "internal scaffolding" between Make and "compile.sh"...
... so that the compiler command is passed down with a CLI arg. that uses a "--compiler_command=" prefix.
This prefix is there to disambiguate the compiler command from the compiler arg.s; this "fixes" the
[admittedly pedantically testing] starting point of "pmake CXX=" [i.e. empty value for CXX],
since Pmake is setting CXXFLAGS to "-O" by default, and without this fix --
i.e. before I added the disambiguation prefix -- "compile.sh" was getting confused by what Pmake gave it
when using this pedantic-test-case starting point, since it "looked" to "compile.sh" like "-O" was the
desired _command basename_ of the _compiler_.  With that value for the alleged compiler command in place,
"compile.sh" [at least on Debian 7] would spew "Illegal option -O" b/c the "which" command was invoked
as "which -O", and "which" [at least on Debian 7] doesn`t accept an "-O" arg./flag/param.